requirements.txt
boto3==1.34.34
python-dotenv==1.0.0
Pillow==10.2.0
pdf2image==1.17.0
gdown==5.1.0
